Added the User class, updated others
Seriously considering using linked lists/nodes/something as a way to create the Batch classes. Will require reconfiguring all Batch classes. perhaps create a Batch class using nodes to connect the tests. Need to also consider how tests are initialized. Is it when the sample is created, or when the test batch is created?